+++ package.json
@@ -40,6 +40,7 @@
     "postcss-loader": "^3.0.0",
     "prop-types": "^15.7.2",
     "react": "^16.10.2",
+    "react-aspect-ratio": "^1.0.41",
     "react-dom": "^16.10.2",
     "react-dropzone-uploader": "^2.10.1",
     "react-modal": "^3.10.1",
