+++ react/components/SideMenu/Background.js
