+++ react/components/SideMenu/Text.js
