+++ react/components/SideMenu/Transition.js
